Automatic even spreading of multiple elements within each of the three battery sections.
For a battery graphic with its three different sections (anode, cathode, and electrolyte), I added code to receive the array of element graphics and spread them accordingly using transforms left and right.

The getTransformStylesForSpread function is a little clunky. I don't really like that it adds one to the middleIndex for even numbers of elements. Consider refactoring that later.

Todos:
1. Convert existing Javascript files to Typescript.
2. Add type annotations to existing code.
3. Write docstrings for existing components.
4. Start writing some type of unit tests. The sooner the better.